Safari-609.1.11.1/LayoutTests/editing/pasteboard/entries-api/datatransfer-items-drop-hidden-file-expected.txt

Test that hidden files are not exposed via Entries API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataTransfer.items.length is 1
PASS directoryEntry.name is "testFiles"
PASS directoryEntry.fullPath is "/testFiles"
PASS directoryEntry.isDirectory is true
PASS foundEntries.length is 4
PASS ex.name is "NotFoundError"
PASS ex.name is "NotFoundError"
PASS successfullyParsed is true

TEST COMPLETE
